Port to Windows
There are still some warnings as well as some asserts. But basically it
works!

+++ include/connection.h
@@ -119,7 +119,7 @@ public:
      *  @param  size        size of the buffer to decode
      *  @return             number of bytes that were processed
      */
-    uint64_t parse(const char *buffer, size_t size)
+    size_t parse(const char *buffer, size_t size)
     {
         return _implementation.parse(ByteBuffer(buffer, size));
     }
@@ -142,7 +142,7 @@ public:
      *  @param  buffer      buffer to decode
      *  @return             number of bytes that were processed
      */
-    uint64_t parse(const Buffer &buffer)
+    size_t parse(const Buffer &buffer)
     {
         return _implementation.parse(buffer);
     }
